driver: rename docker to docker-container
