+++ cmd/buildx/main.go
@@ -8,7 +8,7 @@ import (
 	"github.com/tonistiigi/buildx/commands"
 	"github.com/tonistiigi/buildx/version"
 
-	_ "github.com/tonistiigi/buildx/driver/docker"
+	_ "github.com/tonistiigi/buildx/driver/docker-container"
 )
 
 func main() {
